Extract SessionManager, ToolExecutor, ContextAnalyzer from DochiViewModel
Split DochiViewModel (1062 -> 503 lines) into focused modules:
- SessionManager (175 lines): wake word, session lifecycle, continuous
  conversation, end-session keywords, user identification
- ToolExecutor (132 lines): tool execution loop, image URL extraction
- ContextAnalyzer (266 lines): conversation analysis, memory updates,
  context compression, simple LLM calls

DochiViewModel remains the coordinator: owns all services, wires
callbacks, manages state transitions, and exposes @Published properties
for SwiftUI bindings. Modules access shared state via weak ViewModel
reference.

Closes #1
